third_party/blink/web_tests/svg/dynamic-updates/SVGFEFloodElement-svgdom-flood-color-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'flood-color' property of the SVGFEFloodElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.defaultView.getComputedStyle(floodElement).getPropertyValue('flood-color') is "rgb(145, 32, 103)"
PASS document.defaultView.getComputedStyle(floodElement).getPropertyValue('flood-color') is "rgb(64, 128, 103)"
PASS successfullyParsed is true

TEST COMPLETE
